Create remove_auth.sh
The file remove auth-user-pass line from all the .ovpn file in the current folder. 
This is used because in the script https://github.com/gverni/mytoolbox/blob/master/odroid/etc/init.d/openvpn i add the --auth-user-pass argument to use the up file. 
I could have add the up file into the ovpn, but it's easier to remove the line (using sed) than to add something ad the end of a line.